add configuration file
